requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
aiomysql==0.0.21
aioredis==2.0.0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
cheroot==8.5.2
CherryPy==18.6.1
decorator==5.0.9
environs==9.3.3
flake8==3.9.2
future==0.18.2
idna==3.2
jaraco.classes==3.2.1
jaraco.collections==3.3.0
jaraco.functools==3.3.0
jaraco.text==3.5.1
loguru==0.5.3
lxml==4.6.3
marshmallow==3.13.0
mccabe==0.6.1
more-itertools==8.8.0
multidict==5.1.0
portend==2.7.1
pycodestyle==2.7.0
PyContracts==1.8.12
pyflakes==2.3.1
PyMySQL==0.9.3
pyparsing==2.4.7
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
tempora==4.1.1
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
zc.lockfile==2.0
